+++ build.sh
@@ -119,8 +119,9 @@ LC_UNAME=`echo ${UNAME} | tr '[:upper:]' '[:lower:]'`
 function build_clean {
     echo "Cleaning build directories..."
     rm -Rf out
-    rm -Rf android/filament-android/build
-    rm -Rf android/filament-android/.externalNativeBuild
+    rm -Rf android/filament-android/build android/filament-android/.externalNativeBuild
+    rm -Rf android/filamat-android/build android/filamat-android/.externalNativeBuild
+    rm -Rf android/gltfio-android/build android/gltfio-android/.externalNativeBuild
 }
 
 function build_desktop_target {
@@ -415,7 +416,8 @@ function build_android {
     cd android/gltfio-android
 
     if [[ "$ISSUE_DEBUG_BUILD" == "true" ]]; then
-        ./gradlew -Pfilament_dist_dir=../../out/android-debug/filament assembleDebug
+        ./gradlew -Pfilament_dist_dir=../../out/android-debug/filament assembleDebug \
+                    -Pextra_cmake_args=${VULKAN_ANDROID_OPTION}
 
         if [[ "$INSTALL_COMMAND" ]]; then
             echo "Installing out/gltfio-android-debug.aar..."
@@ -424,7 +426,8 @@ function build_android {
     fi
 
     if [[ "$ISSUE_RELEASE_BUILD" == "true" ]]; then
-        ./gradlew -Pfilament_dist_dir=../../out/android-release/filament assembleRelease
+        ./gradlew -Pfilament_dist_dir=../../out/android-release/filament assembleRelease \
+                -Pextra_cmake_args=${VULKAN_ANDROID_OPTION}
 
         if [[ "$INSTALL_COMMAND" ]]; then
             echo "Installing out/gltfio-android-release.aar..."