+++ filament/backend/src/opengl/OpenGLDriver.cpp
@@ -227,6 +227,12 @@ OpenGLDriver::OpenGLDriver(OpenGLPlatform* platform) noexcept
     disable(GL_DITHER);
     enable(GL_DEPTH_TEST);
 
+    // With desktop GL, the application must enable point size to allow vertex shaders to set it,
+    // but with OpenGL ES, this is always on and there is no enable flag.
+#if GL41_HEADERS
+    enable(GL_PROGRAM_POINT_SIZE);
+#endif
+
     // TODO: Don't enable scissor when it is not necessary. This optimization could be done here in
     // the driver by simply deferring the enable until the scissor rect is smaller than the window.
     enable(GL_SCISSOR_TEST);
@@ -755,10 +761,10 @@ default_case:
 //    GLRenderTarget            : 56        few
 // -- less than 64 bytes
 
-//    GLVertexBuffer            : 80        moderate
+//    GLVertexBuffer            : 208       moderate
 //    GLStream                  : 120       few
 //    GLUniformBuffer           : 128       many
-// -- less than 128 bytes
+// -- less than or equal to 208 bytes
 
 
 OpenGLDriver::HandleAllocator::HandleAllocator(const utils::HeapArea& area)
@@ -812,7 +818,7 @@ template<typename D, typename B, typename ... ARGS>
 typename std::enable_if<std::is_base_of<B, D>::value, D>::type*
 OpenGLDriver::construct(Handle<B> const& handle, ARGS&& ... args) noexcept {
     assert(handle);
-    static_assert(sizeof(D) <= 128, "Handle<> too large");
+    static_assert(sizeof(D) <= 208, "Handle<> too large");
     D* addr = handle_cast<D *>(const_cast<Handle<B>&>(handle));
     new(addr) D(std::forward<ARGS>(args)...);
 #if !defined(NDEBUG) && UTILS_HAS_RTTI
@@ -1022,18 +1028,17 @@ void OpenGLDriver::createTextureR(Handle<HwTexture> th, SamplerType target, uint
 
     GLTexture* t = construct<GLTexture>(th, target, levels, samples, w, h, depth, format, usage);
     if (UTILS_LIKELY(usage & TextureUsage::SAMPLEABLE)) {
-        glGenTextures(1, &t->gl.id);
 
         // below we're using the a = foo(b = C) pattern, this is on purpose, to make sure
         // we don't forget to update targetIndex, and that we do it with the correct value.
         // We DO NOT update targetIndex at function exit to take advantage of the fact that
         // getIndexForTextureTarget() is constexpr -- so all of this disappears at compile time.
 
-        if (UTILS_UNLIKELY(t->target == SamplerType::SAMPLER_EXTERNAL &&
-                           ext.OES_EGL_image_external_essl3)) {
-            t->gl.targetIndex = (uint8_t)
-                    getIndexForTextureTarget(t->gl.target = GL_TEXTURE_EXTERNAL_OES);
+        if (UTILS_UNLIKELY(t->target == SamplerType::SAMPLER_EXTERNAL)) {
+            mPlatform.createExternalImageTexture(t);
         } else {
+            glGenTextures(1, &t->gl.id);
+
             t->gl.internalFormat = getInternalFormat(format);
             assert(t->gl.internalFormat);
 
@@ -1518,7 +1523,11 @@ void OpenGLDriver::destroyTexture(Handle<HwTexture> th) {
             if (t->gl.rb) {
                 glDeleteRenderbuffers(1, &t->gl.rb);
             }
-            glDeleteTextures(1, &t->gl.id);
+            if (UTILS_UNLIKELY(t->target == SamplerType::SAMPLER_EXTERNAL)) {
+                mPlatform.destroyExternalImage(t);
+            } else {
+                glDeleteTextures(1, &t->gl.id);
+            }
         } else {
             assert(t->gl.target == GL_RENDERBUFFER);
             assert(t->gl.rb == 0);
@@ -2092,16 +2101,21 @@ void OpenGLDriver::setCompressedTextureData(GLTexture* t,
 }
 
 void OpenGLDriver::setupExternalImage(void* image) {
+    mPlatform.retainExternalImage(image);
 }
 
 void OpenGLDriver::cancelExternalImage(void* image) {
+    mPlatform.releaseExternalImage(image);
 }
 
 void OpenGLDriver::setExternalImage(Handle<HwTexture> th, void* image) {
+    GLTexture* t = handle_cast<GLTexture*>(th);
+
+    mPlatform.setExternalImage(image, t);
+
+    // TODO: move this logic to PlatformEGL.
     if (ext.OES_EGL_image_external_essl3) {
         DEBUG_MARKER()
-
-        GLTexture* t = handle_cast<GLTexture*>(th);
 
         assert(t->target == SamplerType::SAMPLER_EXTERNAL);
         assert(t->gl.target == GL_TEXTURE_EXTERNAL_OES);