+++ filament/src/Engine.cpp
@@ -202,8 +202,8 @@ void FEngine::init() {
             .build(*this));
     static uint32_t pixel = 0;
     Texture::PixelBufferDescriptor buffer(
-            &pixel, 4, // 4 bytes in 1 RGB pixel
-            Texture::Format::RGB, Texture::Type::UBYTE);
+            &pixel, 4, // 4 bytes in 1 RGBA pixel
+            Texture::Format::RGBA, Texture::Type::UBYTE);
     Texture::FaceOffsets offsets = {};
     mDefaultIblTexture->setImage(*this, 0, std::move(buffer), offsets);
 
@@ -802,6 +802,10 @@ Engine* Engine::create(Backend backend, Platform* platform, void* sharedGLContex
     Engine* handle = engine.get();
     sEngines[handle] = std::move(engine);
     return handle;
+}
+
+void Engine::destroy(Engine* engine) {
+    destroy(&engine);
 }
 
 void Engine::destroy(Engine** engine) {